Following changes are made as per this commit :
1. Added code under base.html, to render Create Event link only for
Organization logged in users, rather than for anyone. This is in regards
to the enhancement that only Organization Users can create events and is
in direct relation to issue #95.
2. Added permission_denied.html template and it's corresponding URL in
urls.py. This is the page rendered when a user goes to event upload link
directly, if he has no access.
3. Changed render condition for Event Upload class under views.py, to
only render for logged in organizer users. Else, it will redirect to
permission denied page.
4. Removed unused html code from event_upload.html template.

Tested this to be working for anonymous users, attendees and Organizer
users. Fixes #104